+++ Classes/Engine2D/Structures/JumpEstructure.hpp
@@ -15,19 +15,19 @@ namespace Structures {
     class JumpEstructure: public BaseStructure {
     private:
         /** @short Default height of the structure */
-        const static unsigned int K_DEFAULT_HEIGHT = 3;
+        const static unsigned short int K_DEFAULT_HEIGHT = 3;
 
         /** @short Default width of the structure */
-        const static unsigned int K_DEFAULT_WIDTH = 20;
+        const static unsigned short int K_DEFAULT_WIDTH = 20;
 
         /** @short Default max value of x jump of the block structure */
-        const static unsigned int K_DEFAULT_MAX_X_DISTANCE_BLOCKS = 2;
+        const static unsigned short int K_DEFAULT_MAX_X_DISTANCE_BLOCKS = 2;
 
         /** @short Default max value of y jump of the block structure */
-        const static unsigned int K_DEFAULT_MAX_Y_DISTANCE_BLOCKS = 1;
+        const static unsigned short int K_DEFAULT_MAX_Y_DISTANCE_BLOCKS = 1;
 
-        unsigned int  _maxXDisBlock;
-        unsigned int  _maxYDisBlock;
+        unsigned short int  _maxXDisBlock;
+        unsigned short int  _maxYDisBlock;
 
 
 
@@ -36,7 +36,7 @@ namespace Structures {
         /** @short Creates a new jump structure with default values */
         JumpEstructure();
         /** @short Creates a new jump structure with parameters if parameteres are not ok, get default values  */
-        JumpEstructure(const unsigned int height, const unsigned int width,const unsigned int max_distanceX,const unsigned int max_distanceY);
+        JumpEstructure(const unsigned short int height, const unsigned  short int width,const unsigned short int max_distanceX,const unsigned short int max_distanceY);
 
     };
 }
